cloud-functions/usage-alerts/requirements.txt
functions-framework==3.*
google-cloud-bigquery==3.*
pandas==2.*
google-cloud-logging==3.*
db-dtypes==1.*